Output + file merging
